Added folders
Optimized sprites, by having a global sprite dictionary, instead of loading a new sprite for each instance. Changed way the buttons are setup in the editor, now allowing for buttons to be created individually, or in a range. A Body2d class is now the base for most classes that it would be relevant for.